+++ CHANGELOG.md
@@ -19,6 +19,7 @@ CHANGELOG
 - Fix validation of parameter `SharedStorage/EfsSettings`: now validation fails when `FileSystemId` is specified 
   along with other `SharedStorage/EfsSettings` parameters, whereas it was previously ignoring them.
 - Fix cluster update when changing the order of SharedStorage together with other changes in the configuration.
+- Avoid failing on DescribeCluster when cluster configuration is not available.
 
 3.2.0
 ------